requirements.txt
numpy==2.0.2
pandas==2.3.0
plotly==6.1.2
sqlalchemy==1.4.26
pyodbc==5.2.0
python-dotenv==0.19.1
urllib3==2.4.0
Flask==2.2.2
Werkzeug==2.2.2